+++ components/plurals/benches/parser.rs
@@ -8,7 +8,6 @@ mod helpers;
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 use icu_provider::prelude::*;
-use icu_provider::structs;
 use std::borrow::Cow;
 
 fn parser(c: &mut Criterion) {
@@ -21,10 +20,10 @@ fn parser(c: &mut Criterion) {
     let mut rules = vec![];
 
     for langid in &plurals_data.langs {
-        let plurals_data: Cow<structs::plurals::PluralRuleStringsV1> = provider
+        let plurals_data: Cow<icu_plurals::provider::PluralRuleStringsV1> = provider
             .load_payload(&DataRequest {
                 resource_path: ResourcePath {
-                    key: structs::plurals::key::CARDINAL_V1,
+                    key: icu_plurals::provider::key::CARDINAL_V1,
                     options: ResourceOptions {
                         variant: None,
                         langid: Some(langid.clone()),
