+++ components/provider/Cargo.toml
@@ -19,15 +19,10 @@ include = [
     "README.md"
 ]
 
-[package.metadata.cargo-all-features]
-skip_optional_dependencies = true
-
 [features]
-# Serialize None values when exporting. Required for Bincode.
-serialize_none = []
-
-# Include the HelloWorldProvider, useful for unit testing.
-hello_world = ["icu_locid_macros"]
+# Include the ErasedDataProvider, useful for type-agnostic data processing.
+# Also required for DataExporter.
+erased = ["serde"]
 
 [dependencies]
 # The serde feature in icu_locid is required for likely subtags support
@@ -36,7 +31,7 @@ tinystr = "0.4"
 erased-serde = "0.3"
 smallstr = { version = "0.2", features = ["serde"] }
 downcast-rs = "1.2"
-serde = { version = "1.0", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"], optional = true }
 writeable = { version = "0.2", path = "../../utils/writeable" }
 
 # For feature "hello_world"
@@ -46,7 +41,12 @@ icu_locid_macros = { version = "0.1", path = "../locid/macros", optional = true 
 serde_json = "1.0"
 static_assertions = "1.1"
 icu_locid_macros = { version = "0.1", path = "../locid/macros" }
+icu_plurals = { version = "0.1", path = "../plurals" }
 
 [[test]]
-name = "iteration"
-required-features = ["hello_world"]
+name = "data_provider"
+required-features = ["erased", "serde"]
+
+[[test]]
+name = "data_receiver"
+required-features = ["erased"]
