--- components/provider/src/structs/mod.rs
@@ -1,8 +1,0 @@
-// This file is part of ICU4X. For terms of use, please see the file
-// called LICENSE at the top level of the ICU4X source tree
-// (online at: https://github.com/unicode-org/icu4x/blob/master/LICENSE ).
-//! Collection of data structures for `DataProvider`.
-pub mod dates;
-pub mod icu4x;
-pub mod likelysubtags;
-pub mod plurals;
