+++ components/provider_cldr/Cargo.toml
@@ -27,9 +27,10 @@ extra_features = [
 ]
 
 [dependencies]
-icu_provider = { version = "0.1", path = "../provider" }
+icu_provider = { version = "0.1", path = "../provider", features = ["erased"] }
 icu_locid = { version = "0.1", path = "../locid" }
 icu_plurals = { version = "0.1", path = "../plurals" }
+icu_datetime = { version = "0.1", path = "../datetime" }
 json = "0.12"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
@@ -48,5 +49,11 @@ mktemp = "0.4"
 icu_locid_macros = { version = "0.1", path = "../locid/macros" }
 
 [features]
+# Serialize None values when exporting. Required for Bincode.
+serialize_none = [
+    "icu_datetime/serialize_none",
+    "icu_plurals/serialize_none",
+]
+
 # Automatically download CLDR data from a host
 download = ["urlencoding", "reqwest", "unzip", "dirs", "log"]
