+++ components/provider_cldr/src/transform/dates.rs
@@ -5,8 +5,8 @@ use crate::cldr_langid::CldrLangID;
 use crate::error::Error;
 use crate::reader::{get_subdirectories, open_reader};
 use crate::CldrPaths;
+use icu_datetime::provider::*;
 use icu_provider::prelude::*;
-use icu_provider::structs::dates::*;
 use std::borrow::Cow;
 use std::convert::TryFrom;
 use std::marker::PhantomData;
