+++ components/provider_fs/tests/test_file_io.rs
@@ -4,43 +4,41 @@
 
 use icu_locid::LanguageIdentifier;
 use icu_locid_macros::langid;
+use icu_plurals::provider::*;
 use icu_provider::erased::*;
 use icu_provider::prelude::*;
-use icu_provider::structs;
 use icu_provider_fs::FsDataProvider;
 use std::borrow::Cow;
 
-const EXPECTED_RU_DATA: structs::plurals::PluralRuleStringsV1 =
-    structs::plurals::PluralRuleStringsV1 {
-        zero: None,
-        one: Some(Cow::Borrowed("v = 0 and i % 10 = 1 and i % 100 != 11")),
-        two: None,
-        few: Some(Cow::Borrowed(
-            "v = 0 and i % 10 = 2..4 and i % 100 != 12..14",
-        )),
-        many: Some(Cow::Borrowed(
-            "v = 0 and i % 10 = 0 or v = 0 and i % 10 = 5..9 or v = 0 and i % 100 = 11..14",
-        )),
-    };
+const EXPECTED_RU_DATA: PluralRuleStringsV1 = PluralRuleStringsV1 {
+    zero: None,
+    one: Some(Cow::Borrowed("v = 0 and i % 10 = 1 and i % 100 != 11")),
+    two: None,
+    few: Some(Cow::Borrowed(
+        "v = 0 and i % 10 = 2..4 and i % 100 != 12..14",
+    )),
+    many: Some(Cow::Borrowed(
+        "v = 0 and i % 10 = 0 or v = 0 and i % 10 = 5..9 or v = 0 and i % 100 = 11..14",
+    )),
+};
 
 #[cfg(feature = "bincode")]
-const EXPECTED_SR_DATA: structs::plurals::PluralRuleStringsV1 =
-    structs::plurals::PluralRuleStringsV1 {
-        zero: None,
-        one: Some(Cow::Borrowed(
-            "v = 0 and i % 10 = 1 and i % 100 != 11 or f % 10 = 1 and f % 100 != 11",
-        )),
-        two: None,
-        few: Some(Cow::Borrowed(
-            "v = 0 and i % 10 = 2..4 and i % 100 != 12..14 or f % 10 = 2..4 and f % 100 != 12..14",
-        )),
-        many: None,
-    };
+const EXPECTED_SR_DATA: PluralRuleStringsV1 = PluralRuleStringsV1 {
+    zero: None,
+    one: Some(Cow::Borrowed(
+        "v = 0 and i % 10 = 1 and i % 100 != 11 or f % 10 = 1 and f % 100 != 11",
+    )),
+    two: None,
+    few: Some(Cow::Borrowed(
+        "v = 0 and i % 10 = 2..4 and i % 100 != 12..14 or f % 10 = 2..4 and f % 100 != 12..14",
+    )),
+    many: None,
+};
 
 fn get_request(langid: LanguageIdentifier) -> DataRequest {
     DataRequest {
         resource_path: ResourcePath {
-            key: structs::plurals::key::CARDINAL_V1,
+            key: key::CARDINAL_V1,
             options: ResourceOptions {
                 variant: None,
                 langid: Some(langid),
@@ -54,7 +52,7 @@ fn test_json() {
     let provider = FsDataProvider::try_new("./tests/testdata/json")
         .expect("Loading file from testdata directory");
 
-    let plurals_data: Cow<structs::plurals::PluralRuleStringsV1> = provider
+    let plurals_data: Cow<PluralRuleStringsV1> = provider
         .load_payload(&get_request(langid!("ru")))
         .expect("The data should be valid")
         .take_payload()
@@ -67,8 +65,7 @@ fn test_json_dyn_erased() {
     let provider = FsDataProvider::try_new("./tests/testdata/json")
         .expect("Loading file from testdata directory");
 
-    let plurals_data: Cow<structs::plurals::PluralRuleStringsV1> = (&provider
-        as &dyn ErasedDataProvider)
+    let plurals_data: Cow<PluralRuleStringsV1> = (&provider as &dyn ErasedDataProvider)
         .load_payload(&get_request(langid!("ru")))
         .expect("The data should be valid")
         .take_payload()
@@ -80,13 +77,13 @@ fn test_json_dyn_erased() {
 fn test_json_errors() {
     let provider = FsDataProvider::try_new("./tests/testdata/json")
         .expect("Loading file from testdata directory");
-    let mut receiver = DataReceiver::<structs::plurals::PluralRuleStringsV1>::new();
+    let mut receiver = DataReceiver::<PluralRuleStringsV1>::new();
 
     assert!(matches!(
         provider.load_to_receiver(
             &DataRequest {
                 resource_path: ResourcePath {
-                    key: structs::plurals::key::CARDINAL_V1,
+                    key: key::CARDINAL_V1,
                     options: ResourceOptions {
                         variant: None,
                         langid: Some(langid!("ru"))
@@ -104,7 +101,7 @@ fn test_json_errors() {
         provider.load_to_receiver(
             &DataRequest {
                 resource_path: ResourcePath {
-                    key: structs::plurals::key::CARDINAL_V1,
+                    key: key::CARDINAL_V1,
                     options: ResourceOptions {
                         variant: None,
                         langid: Some(langid!("zh"))
@@ -122,7 +119,7 @@ fn test_json_errors() {
         provider.load_to_receiver(
             &DataRequest {
                 resource_path: ResourcePath {
-                    key: structs::plurals::key::ORDINAL_V1,
+                    key: key::ORDINAL_V1,
                     options: ResourceOptions {
                         variant: None,
                         langid: Some(langid!("ru"))
@@ -140,7 +137,7 @@ fn test_json_errors() {
         provider.load_to_receiver(
             &DataRequest {
                 resource_path: ResourcePath {
-                    key: structs::dates::key::GREGORY_V1,
+                    key: icu_provider::hello_world::key::HELLO_WORLD_V1,
                     options: ResourceOptions {
                         variant: None,
                         langid: Some(langid!("ru"))
@@ -154,12 +151,18 @@ fn test_json_errors() {
 }
 
 #[test]
+#[cfg(feature = "export")]
+fn test_json_export() {
+    //
+}
+
+#[test]
 #[cfg(feature = "bincode")]
 fn test_bincode() {
     let provider = FsDataProvider::try_new("./tests/testdata/bincode")
         .expect("Loading file from testdata directory");
 
-    let plurals_data: Cow<structs::plurals::PluralRuleStringsV1> = provider
+    let plurals_data: Cow<PluralRuleStringsV1> = provider
         .load_payload(&get_request(langid!("sr")))
         .expect("The data should be valid")
         .take_payload()
@@ -173,8 +176,7 @@ fn test_bincode_dyn_erased() {
     let provider = FsDataProvider::try_new("./tests/testdata/bincode")
         .expect("Loading file from testdata directory");
 
-    let plurals_data: Cow<structs::plurals::PluralRuleStringsV1> = (&provider
-        as &dyn ErasedDataProvider)
+    let plurals_data: Cow<PluralRuleStringsV1> = (&provider as &dyn ErasedDataProvider)
         .load_payload(&get_request(langid!("sr")))
         .expect("The data should be valid")
         .take_payload()
